orchestrator/clickhouse: buffer networks.csv to disk
And build it on changes, not on the fly.
